Questions
Q1. What is Abstraction in OOps? Explain with an example.
Q2. Differentiate between Abstraction and Encapsulation. Explain with an example.
Q3. What is abc module in python? Why is it used?
Q4. How can we achieve data abstraction?
Q5. Can we create an instance of an abstract class? Explain your answer.