Questions
Q1. You are writing code for a company. The requirement of the company is that you create a python
function that will check whether the password entered by the user is correct or not. The function should
take the password as input and return the string “Valid Password” if the entered password follows the
below-given password guidelines else it should return “Invalid Password”.
Note: 1. The Password should contain at least two uppercase letters and at least two lowercase letters.
2. The Password should contain at least a number and three special characters.
3. The length of the password should be 10 characters long.
Q2. Solve the below-given questions using at least one of the following:
1. Lambda functioJ
2. Filter functioJ
3. Zap functioJ
4. List ComprehensioI
B Check if the string starts with a particular letterY
B Check if the string is numericY
B Sort a list of tuples having fruit names and their quantity. [("mango",99),("orange",80), ("grapes", 1000)-
B Find the squares of numbers from 1 to 10Y
B Find the cube root of numbers from 1 to 10Y
B Check if a given number is evenY
B Filter odd numbers from the given list.
[1,2,3,4,5,6,7,8,9,10-
B Sort a list of integers into positive and negative integers lists.
[1,2,3,4,5,6,-1,-2,-3,-4,-5,0]